Implement improved GUI for ChaosGame and ExploreGame
To display explore Julia sets better and faster a IntStream was implemented to compute the sets faster.
The boundaries in the while-loop was edited to work correctly. And the canvas int array was changed to a
double array. A few more methods was added to Vector2D.
Buttons to zoom and the ability to drag and move the fractal was added to ExplorePage.
A ComboBox was added to ChaosPage giving the ability to select different fractals to run as a Chaos Game.

+++ src/main/java/org/example/chaosgame/chaos/ChaosCanvas.java
@@ -17,7 +17,7 @@ import org.example.chaosgame.transformations.AffineTransform2D;
 public class ChaosCanvas {
   private final int width;
   private final int height;
-  private final int[][] canvas;
+  private final double[][] canvas;
   private final Vector2D minCoords;
   private final Vector2D maxCoords;
   private final AffineTransform2D transformCoordsToIndices;
@@ -54,10 +54,10 @@ public class ChaosCanvas {
                     / (maxCoords.getY() - minCoords.getY())),
                     ((width - 1.0) * minCoords.getX()) / (minCoords.getX() - maxCoords.getX())
   ));
-    this.canvas = new int[height][width];
+    this.canvas = new double[height][width];
   }
 
-  public int getPixel(Vector2D point){
+  public double getPixel(Vector2D point){
     Vector2D indices = transformCoordsToIndices.transform(point);
     int x = (int) indices.getX();
     int y = (int) indices.getY();
@@ -89,7 +89,7 @@ public class ChaosCanvas {
    *
    * @param iter The iteration value to set the pixel to
    */
-  public void putPixel(int x, int y, int iter){
+  public void putPixel(int x, int y, double iter){
     if (y >= 0 && y < height && x >= 0 && x < width) {
       canvas[y][x] = iter;
     }
@@ -113,7 +113,7 @@ public class ChaosCanvas {
     return new Vector2D(x, y);
   }
 
-  public int[][] getCanvasArray() {
+  public double[][] getCanvasArray() {
     return canvas;
   }
